pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amtsguide-readiness-spec"
version = "0.1.0"
description = "Schemas and validators for reviewable AI work products"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AmtsGuide", email = "dev@amtsguide.de"}
]
keywords = ["ai", "validation", "schemas", "government", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
validate-work-product = "amtsguide_readiness_spec.cli:validate_work_product"
validate-text = "amtsguide_readiness_spec.cli:validate_text"

[project.urls]
Homepage = "https://github.com/andxblink/amtsguide_public"
Repository = "https://github.com/andxblink/amtsguide_public"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

